Move bytecounter to iostream
